The app is now fully operational for production deployment
works with Vercel
Had to upgrade how the JSONs are loaded by creating a combination of precomplilation scripts to include them, since there is no folder structure in production environments.

This also means the app will be much more dApp ready!

(Next: make it a dApp and deploy it on WTTP)